new playgame function refactor

+++ src/main/kotlin/org/wit/cm/models/ScoreModel.kt
@@ -1,1 +1,1 @@
-data class ScoreModel(var userName: String ="", var score: Int=0)
+data class ScoreModel(var userName: String ="", var score: Long=0)
